requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.3
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
idna==3.4
inflection==0.5.1
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.2
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.15
